LMS-2173: Move to the appropriate scheduled task to be able to filter successful events.
